backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pillow==10.2.0
python-dotenv==1.0.0
openai==1.12.0
aiofiles==23.2.1
bcrypt==4.1.2
